# Conflicts:
#   src/test/java/tests/CreateOrderTest.java

+++ src/test/java/models/order/TotalAmount.java
@@ -7,11 +7,11 @@ import lombok.Data;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public @Data class TotalAmount {
     @JsonProperty("RUB")
-    public double rUB;
+    public int rUB;
     @JsonProperty("USD")
     public double uSD;
     @JsonProperty("EUR")
     public double eUR;
     @JsonProperty("KZT")
-    public double kZT;
+    public int kZT;
 }
